[#29] Split slack code into correct client and routes file
